Refactor SaveLoadMixin used by both Wrapper and Results
save and load methods are essentially identical in ResultsWrapper and LikelihoodModelResults.  Moving those into a single SaveLoadMixin
de-duplicates code, reduces number of points-of-contact for model.data object
