+++ statsmodels/base/model.py
@@ -844,7 +844,9 @@ class Results(object):
 
 
 #TODO: public method?
-class LikelihoodModelResults(Results):
+# TODO: is it redundant to have Results inherit SaveLoadMixin given that
+#   ResultsWrapper already inherits it?
+class LikelihoodModelResults(Results, wrap.SaveLoadMixin):
     """
     Class to contain results from likelihood models
 
@@ -1181,8 +1183,7 @@ class LikelihoodModelResults(Results):
             return cov_p
 
     #TODO: make sure this works as needed for GLMs
-    def t_test(self, r_matrix, cov_p=None, scale=None,
-               use_t=None):
+    def t_test(self, r_matrix, cov_p=None, scale=None, use_t=None):
         """
         Compute a t-test for a each linear hypothesis of the form Rb = q
 
@@ -1733,111 +1734,7 @@ class LikelihoodModelResults(Results):
             upper = self.params[cols] + q * bse[cols]
         return np.asarray(lzip(lower, upper))
 
-    def save(self, fname, remove_data=False):
-        '''
-        save a pickle of this instance
 
-        Parameters
-        ----------
-        fname : string or filehandle
-            fname can be a string to a file path or filename, or a filehandle.
-        remove_data : bool
-            If False (default), then the instance is pickled without changes.
-            If True, then all arrays with length nobs are set to None before
-            pickling. See the remove_data method.
-            In some cases not all arrays will be set to None.
-
-        Notes
-        -----
-        If remove_data is true and the model result does not implement a
-        remove_data method then this will raise an exception.
-
-        '''
-
-        from statsmodels.iolib.smpickle import save_pickle
-
-        if remove_data:
-            self.remove_data()
-
-        save_pickle(self, fname)
-
-    @classmethod
-    def load(cls, fname):
-        '''
-        load a pickle, (class method)
-
-        Parameters
-        ----------
-        fname : string or filehandle
-            fname can be a string to a file path or filename, or a filehandle.
-
-        Returns
-        -------
-        unpickled instance
-
-        '''
-
-        from statsmodels.iolib.smpickle import load_pickle
-        return load_pickle(fname)
-
-    def remove_data(self):
-        '''remove data arrays, all nobs arrays from result and model
-
-        This reduces the size of the instance, so it can be pickled with less
-        memory. Currently tested for use with predict from an unpickled
-        results and model instance.
-
-        .. warning:: Since data and some intermediate results have been removed
-           calculating new statistics that require them will raise exceptions.
-           The exception will occur the first time an attribute is accessed
-           that has been set to None.
-
-        Not fully tested for time series models, tsa, and might delete too much
-        for prediction or not all that would be possible.
-
-        The lists of arrays to delete are maintained as attributes of
-        the result and model instance, except for cached values. These
-        lists could be changed before calling remove_data.
-
-        The attributes to remove are named in:
-
-        model._data_attr : arrays attached to both the model instance
-            and the results instance with the same attribute name.
-
-        result.data_in_cache : arrays that may exist as values in
-            result._cache (TODO : should privatize name)
-
-        result._data_attr_model : arrays attached to the model
-            instance but not to the results instance
-        '''
-        def wipe(obj, att):
-            #get to last element in attribute path
-            p = att.split('.')
-            att_ = p.pop(-1)
-            try:
-                obj_ = reduce(getattr, [obj] + p)
-
-                #print(repr(obj), repr(att))
-                #print(hasattr(obj_, att_))
-                if hasattr(obj_, att_):
-                    #print('removing3', att_)
-                    setattr(obj_, att_, None)
-            except AttributeError:
-                pass
-
-        model_only = ['model.' + i for i in getattr(self, "_data_attr_model", [])]
-        model_attr = ['model.' + i for i in self.model._data_attr]
-        for att in self._data_attr + model_attr + model_only:
-            #print('removing', att)
-            wipe(self, att)
-
-        data_in_cache = getattr(self, 'data_in_cache', [])
-        data_in_cache += ['fittedvalues', 'resid', 'wresid']
-        for key in data_in_cache:
-            try:
-                self._cache[key] = None
-            except (AttributeError, KeyError):
-                pass
 
 
 class LikelihoodResultsWrapper(wrap.ResultsWrapper):
